python/requirements.txt
Flask
azure-ai-textanalytics
Werkzeug

#Flask==2.1.1
#azure-ai-textanalytics==5.2.0b2
#Werkzeug==2.2.2